refactor: add DeleteAccountDialog
